refactor(node/actuator): mv proposal actuators to stand-alone file
